asiby:BUG:Ticket #167: Bug this call monitoring ... ERROR: Cannot download file /var/spool/asterisk/monitor/...
git-svn-id: http://svn.a2billing.net/svn/asterisk2billing/trunk@457 9d51a8f0-6521-42ea-932e-d822aa02ce2f

+++ A2Billing_UI/lib/Misc.php
@@ -246,12 +246,14 @@ function remove_prefix($phonenumber){
  * function linkonmonitorfile
  */
 function linkonmonitorfile($value){
-			  
-   $myfile = $value.".".MONITOR_FORMATFILE;
-   $myfile = base64_encode($myfile);
-   echo "<a target=_blank href=\"call-log-customers.php?download=file&file=".$myfile."\">";
-   echo '<img src="'.Images_Path.'/stock-mic.png" height="18" /></a>';
-   
+	$myfile = $value.".".MONITOR_FORMATFILE;
+	$dl_full = MONITOR_PATH."/".$myfile;
+	if (!file_exists($dl_full)){
+		return;
+	}
+	$myfile = base64_encode($myfile);
+	echo "<a target=_blank href=\"call-log-customers.php?download=file&file=".$myfile."\">";
+	echo '<img src="'.Images_Path.'/stock-mic.png" height="18" /></a>';   
 }
 
 function linktocustomer($value){
